Simulate the queuing system of a bank

This function takes information about customer arrival and service times and simulates the queuing system of a bank.

Multiserver(Arrivals, ServiceTimes, NumServers = 1)

Arguments

Arrivals
a vector of customer arrival times in ascending order

ServiceTimes
a vector of corresponding customer service times

NumServers
an integer of the number of service servers of a bank

Value

a data.frame simulating the queuing system of the bank:

Arrivals

arrival time of a customer, in seconds

ServiceBegins

the time that the customer service begins

ChosenServer

the server that the customer goes through

ServiceEnds

the time that the customer service ends

Examples

Multiserver(bank$arrival_time, bank$service_time, 3)

#>      Arrivals ServiceBegins ChosenServer ServiceEnds
#> 1    118.9688      118.9688            1    148.7928
#> 2    133.9700      133.9700            2    288.9776
#> 3    325.6309      325.6309            3    356.3327
#> 4    338.5996      338.5996            1    424.1283
#> 5    371.9286      371.9286            2    652.2320
#> 6    395.9377      395.9377            3    474.0537
#> 7    399.8443      424.1283            1    459.2263
#> 8    491.0389      491.0389            1    649.8293
#> 9    531.2127      531.2127            3    761.8258
#> 10   627.7234      649.8293            1    756.1308
#> 11   696.9662      696.9662            2   1133.2617
#> 12   791.0388      791.0388            1   1007.2719
#> 13   803.3939      803.3939            3    932.4672
#> 14   808.8096      932.4672            3    967.2119
#> 15   819.9051      967.2119            3   1173.8581
#> 16   820.0413     1007.2719            1   1086.2532
#> 17   920.9695     1086.2532            1   1157.5569
#> 18   926.9460     1133.2617            2   1369.7752
#> 19   931.2104     1157.5569            1   1185.6023
#> 20   946.3734     1173.8581            3   1322.9064
#> 21  1066.6020     1185.6023            1   1255.7955
#> 22  1077.3357     1255.7955            1   1295.9332
#> 23  1084.9992     1295.9332            1   1352.2103
#> 24  1160.8246     1322.9064            3   1353.9990
#> 25  1240.9804     1352.2103            1   1413.2847
#> 26  1251.1413     1353.9990            3   1684.8444
#> 27  1293.3333     1369.7752            2   1397.9424
#> 28  1404.9411     1404.9411            2   1463.8905
#> 29  1420.6269     1420.6269            1   1452.7087
#> 30  1426.5267     1452.7087            1   1554.7481
#> 31  1442.7966     1463.8905            2   1682.2007
#> 32  1463.4056     1554.7481            1   1603.3354
#> 33  1585.9136     1603.3354            1   1631.7283
#> 34  1678.9439     1678.9439            1   2040.1855
#> 35  1716.9316     1716.9316            2   1923.8466
#> 36  1722.5008     1722.5008            3   2078.1926
#> 37  1831.3243     1923.8466            2   1971.1074
#> 38  1895.6954     1971.1074            2   2364.7692
#> 39  1947.3189     2040.1855            1   2119.8011
#> 40  1974.5048     2078.1926            3   2141.4340
#> 41  1991.2259     2119.8011            1   2173.8524
#> 42  2169.9533     2169.9533            3   2304.4491
#> 43  2187.0929     2187.0929            1   2395.1468
#> 44  2302.4389     2304.4491            3   2701.9152
#> 45  2333.8023     2364.7692            2   2567.7615
#> 46  2354.9218     2395.1468            1   2489.8901
#> 47  2422.9174     2489.8901            1   2528.1612
#> 48  2484.2988     2528.1612            1   2585.0378
#> 49  2493.1107     2567.7615            2   2751.3635
#> 50  2518.2996     2585.0378            1   3147.5789
#> 51  2519.9548     2701.9152            3   2834.7392
#> 52  2562.4772     2751.3635            2   2894.7929
#> 53  2581.0792     2834.7392            3   2928.3057
#> 54  2600.4818     2894.7929            2   3142.7728
#> 55  2660.1844     2928.3057            3   3126.4606
#> 56  2716.3124     3126.4606            3   3219.9158
#> 57  2756.0081     3142.7728            2   3444.8928
#> 58  2775.0100     3147.5789            1   3269.6311
#> 59  2798.2185     3219.9158            3   3323.3357
#> 60  2836.1702     3269.6311            1   3724.9807
#> 61  2907.9819     3323.3357            3   3358.6663
#> 62  2997.1539     3358.6663            3   3472.5343
#> 63  3019.5207     3444.8928            2   3562.9134
#> 64  3047.7464     3472.5343            3   3767.0206
#> 65  3192.6444     3562.9134            2   3805.4636
#> 66  3193.2759     3724.9807            1   3820.5113
#> 67  3270.6307     3767.0206            3   3836.9807
#> 68  3353.4185     3805.4636            2   4044.2070
#> 69  3403.6330     3820.5113            1   3890.0381
#> 70  3432.3373     3836.9807            3   3975.0275
#> 71  3454.6138     3890.0381            1   4194.3732
#> 72  3474.4198     3975.0275            3   4291.1735
#> 73  3515.3810     4044.2070            2   4313.5512
#> 74  3522.3617     4194.3732            1   4434.2702
#> 75  3591.1576     4291.1735            3   4351.5070
#> 76  3602.6851     4313.5512            2   4416.5514
#> 77  3625.2998     4351.5070            3   4516.9895
#> 78  3658.8517     4416.5514            2   4741.3108
#> 79  3678.6289     4434.2702            1   4619.8519
#> 80  3781.4736     4516.9895            3   4665.9455
#> 81  3801.8675     4619.8519            1   4849.0462
#> 82  3904.9857     4665.9455            3   5215.2983
#> 83  3951.5178     4741.3108            2   4937.4053
#> 84  4056.6002     4849.0462            1   5878.2614
#> 85  4091.0934     4937.4053            2   5377.1948
#> 86  4204.7864     5215.2983            3   5241.7020
#> 87  4306.8493     5241.7020            3   5283.5628
#> 88  4348.4671     5283.5628            3   5313.9305
#> 89  4441.3592     5313.9305            3   5415.4478
#> 90  4498.8255     5377.1948            2   5765.1661
#> 91  4528.3863     5415.4478            3   5969.6341
#> 92  4552.6861     5765.1661            2   6051.6400
#> 93  4698.6234     5878.2614            1   5959.1253
#> 94  4702.2331     5959.1253            1   6050.9490
#> 95  4744.2914     5969.6341            3   6018.2974
#> 96  4744.4531     6018.2974            3   6101.0153
#> 97  4892.4397     6050.9490            1   6084.6517
#> 98  4947.8026     6051.6400            2   6127.0099
#> 99  4961.8806     6084.6517            1   6270.2034
#> 100 5029.9246     6101.0153            3   6144.1942